+++ src/caroodj/Menus/Panel/Home.java
@@ -4,18 +4,25 @@
  */
 package Menus.Panel;
 
+import java.awt.CardLayout;
+
 /**
  *
- * @author JUSTIN
+ * @author William
  */
 public class Home extends javax.swing.JPanel {
 
     /**
-     * Creates new form Home
+     * Creates new form GUI
      */
-    public Home(String personId) {
-        this.personId = personId;
+    
+    
+    private CardLayout cardLayout;
+    private String personId;
+    public Home(String p) {
+        this.personId = p;
         initComponents();
+        cardLayout = (CardLayout)(panelCard.getLayout());
     }
 
     /**
@@ -27,31 +34,307 @@ public class Home extends javax.swing.JPanel {
     // <editor-fold defaultstate="collapsed" desc="Generated Code">                          
     private void initComponents() {
 
-        jLabel1 = new javax.swing.JLabel();
+        jPanel3 = new javax.swing.JPanel();
+        jLabel51 = new javax.swing.JLabel();
+        jSeparator4 = new javax.swing.JSeparator();
+        jButton3 = new javax.swing.JButton();
+        jButton4 = new javax.swing.JButton();
+        jButton5 = new javax.swing.JButton();
+        jButton6 = new javax.swing.JButton();
+        jButton7 = new javax.swing.JButton();
+        jPanel1 = new javax.swing.JPanel();
+        c14 = new javax.swing.JPanel();
+        jLabel11 = new javax.swing.JLabel();
+        c3 = new javax.swing.JPanel();
+        jLabel12 = new javax.swing.JLabel();
+        c2 = new javax.swing.JPanel();
+        jLabel13 = new javax.swing.JLabel();
+        c1 = new javax.swing.JPanel();
+        jLabel14 = new javax.swing.JLabel();
+        c15 = new javax.swing.JPanel();
+        jLabel15 = new javax.swing.JLabel();
 
-        jLabel1.setText(this.personId);
+        jPanel3.setBackground(new java.awt.Color(49, 74, 86));
+
+        jLabel51.setFont(new java.awt.Font("Tw Cen MT Condensed", 1, 48)); // NOI18N
+        jLabel51.setForeground(new java.awt.Color(255, 255, 255));
+        jLabel51.setText("ZU5COVI");
+
+        jSeparator4.setBackground(new java.awt.Color(255, 255, 255));
+
+        jButton3.setBackground(new java.awt.Color(49, 74, 86));
+        jButton3.setFont(new java.awt.Font("Yu Gothic UI", 1, 24)); // NOI18N
+        jButton3.setForeground(new java.awt.Color(255, 255, 255));
+        jButton3.setText("Account");
+        jButton3.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton3ActionPerformed(evt);
+            }
+        });
+
+        jButton4.setBackground(new java.awt.Color(49, 74, 86));
+        jButton4.setFont(new java.awt.Font("Yu Gothic UI", 1, 24)); // NOI18N
+        jButton4.setForeground(new java.awt.Color(255, 255, 255));
+        jButton4.setText("Rental");
+        jButton4.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton4ActionPerformed(evt);
+            }
+        });
+
+        jButton5.setBackground(new java.awt.Color(49, 74, 86));
+        jButton5.setFont(new java.awt.Font("Yu Gothic UI", 1, 24)); // NOI18N
+        jButton5.setForeground(new java.awt.Color(255, 255, 255));
+        jButton5.setText("Home");
+        jButton5.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton5ActionPerformed(evt);
+            }
+        });
+
+        jButton6.setBackground(new java.awt.Color(49, 74, 86));
+        jButton6.setFont(new java.awt.Font("Yu Gothic UI", 1, 24)); // NOI18N
+        jButton6.setForeground(new java.awt.Color(255, 255, 255));
+        jButton6.setText("Messages");
+        jButton6.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton6ActionPerformed(evt);
+            }
+        });
+
+        jButton7.setBackground(new java.awt.Color(49, 74, 86));
+        jButton7.setFont(new java.awt.Font("Yu Gothic UI", 1, 24)); // NOI18N
+        jButton7.setForeground(new java.awt.Color(255, 255, 255));
+        jButton7.setText("History");
+        jButton7.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jButton7ActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout jPanel3Layout = new javax.swing.GroupLayout(jPanel3);
+        jPanel3.setLayout(jPanel3Layout);
+        jPanel3Layout.setHorizontalGroup(
+            jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel3Layout.createSequentialGroup()
+                .addGap(43, 43, 43)
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addComponent(jSeparator4, javax.swing.GroupLayout.PREFERRED_SIZE, 150, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel51))
+                .addContainerGap(44, Short.MAX_VALUE))
+            .addGroup(jPanel3Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jButton5, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jButton4, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jButton6, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jButton7, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jButton3, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        jPanel3Layout.setVerticalGroup(
+            jPanel3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel3Layout.createSequentialGroup()
+                .addGap(34, 34, 34)
+                .addComponent(jLabel51)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jSeparator4, javax.swing.GroupLayout.PREFERRED_SIZE, 10, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
+                .addComponent(jButton5, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(jButton4, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(jButton6, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(jButton3, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(12, 12, 12)
+                .addComponent(jButton7, javax.swing.GroupLayout.PREFERRED_SIZE, 71, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(18, Short.MAX_VALUE))
+        );
+
+        jPanel1.setLayout(new java.awt.CardLayout());
+
+        c14.setBackground(new java.awt.Color(102, 102, 255));
+
+        jLabel11.setText("acc");
+
+        javax.swing.GroupLayout c14Layout = new javax.swing.GroupLayout(c14);
+        c14.setLayout(c14Layout);
+        c14Layout.setHorizontalGroup(
+            c14Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c14Layout.createSequentialGroup()
+                .addGap(115, 115, 115)
+                .addComponent(jLabel11, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(469, Short.MAX_VALUE))
+        );
+        c14Layout.setVerticalGroup(
+            c14Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c14Layout.createSequentialGroup()
+                .addGap(202, 202, 202)
+                .addComponent(jLabel11)
+                .addContainerGap(317, Short.MAX_VALUE))
+        );
+
+        jPanel1.add(c14, "card4");
+
+        c3.setBackground(new java.awt.Color(255, 204, 204));
+
+        jLabel12.setText("msg");
+
+        javax.swing.GroupLayout c3Layout = new javax.swing.GroupLayout(c3);
+        c3.setLayout(c3Layout);
+        c3Layout.setHorizontalGroup(
+            c3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c3Layout.createSequentialGroup()
+                .addGap(180, 180, 180)
+                .addComponent(jLabel12, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(404, Short.MAX_VALUE))
+        );
+        c3Layout.setVerticalGroup(
+            c3Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c3Layout.createSequentialGroup()
+                .addGap(166, 166, 166)
+                .addComponent(jLabel12)
+                .addContainerGap(353, Short.MAX_VALUE))
+        );
+
+        jPanel1.add(c3, "card3");
+
+        c2.setBackground(new java.awt.Color(255, 204, 0));
+
+        jLabel13.setText("rental");
+
+        javax.swing.GroupLayout c2Layout = new javax.swing.GroupLayout(c2);
+        c2.setLayout(c2Layout);
+        c2Layout.setHorizontalGroup(
+            c2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c2Layout.createSequentialGroup()
+                .addGap(190, 190, 190)
+                .addComponent(jLabel13, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(394, Short.MAX_VALUE))
+        );
+        c2Layout.setVerticalGroup(
+            c2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c2Layout.createSequentialGroup()
+                .addGap(194, 194, 194)
+                .addComponent(jLabel13)
+                .addContainerGap(325, Short.MAX_VALUE))
+        );
+
+        jPanel1.add(c2, "card2");
+
+        c1.setBackground(new java.awt.Color(204, 0, 255));
+
+        jLabel14.setText("home");
+
+        javax.swing.GroupLayout c1Layout = new javax.swing.GroupLayout(c1);
+        c1.setLayout(c1Layout);
+        c1Layout.setHorizontalGroup(
+            c1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c1Layout.createSequentialGroup()
+                .addGap(288, 288, 288)
+                .addComponent(jLabel14, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(296, Short.MAX_VALUE))
+        );
+        c1Layout.setVerticalGroup(
+            c1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c1Layout.createSequentialGroup()
+                .addGap(124, 124, 124)
+                .addComponent(jLabel14)
+                .addContainerGap(395, Short.MAX_VALUE))
+        );
+
+        jPanel1.add(c1, "card1");
+
+        c15.setBackground(new java.awt.Color(102, 255, 204));
+
+        jLabel15.setText("history");
+
+        javax.swing.GroupLayout c15Layout = new javax.swing.GroupLayout(c15);
+        c15.setLayout(c15Layout);
+        c15Layout.setHorizontalGroup(
+            c15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c15Layout.createSequentialGroup()
+                .addGap(228, 228, 228)
+                .addComponent(jLabel15, javax.swing.GroupLayout.PREFERRED_SIZE, 37, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(356, Short.MAX_VALUE))
+        );
+        c15Layout.setVerticalGroup(
+            c15Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(c15Layout.createSequentialGroup()
+                .addGap(190, 190, 190)
+                .addComponent(jLabel15)
+                .addContainerGap(329, Short.MAX_VALUE))
+        );
+
+        jPanel1.add(c15, "card5");
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
         this.setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(176, 176, 176)
-                .addComponent(jLabel1)
-                .addContainerGap(194, Short.MAX_VALUE))
+                .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, 621, Short.MAX_VALUE)
+                .addContainerGap())
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(jPanel3, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
             .addGroup(layout.createSequentialGroup()
-                .addGap(136, 136, 136)
-                .addComponent(jLabel1)
-                .addContainerGap(148, Short.MAX_VALUE))
+                .addContainerGap()
+                .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, 535, Short.MAX_VALUE))
         );
     }// </editor-fold>                        
 
+    private void jButton3ActionPerformed(java.awt.event.ActionEvent evt) {                                         
+        // TODO add your handling code here:
+        cardLayout.show(panelCard, "card4");
+    }                                        
+
+    private void jButton4ActionPerformed(java.awt.event.ActionEvent evt) {                                         
+        // TODO add your handling code here:
+        cardLayout.show(panelCard, "card2");
+    }                                        
+
+    private void jButton5ActionPerformed(java.awt.event.ActionEvent evt) {                                         
+        // TODO add your handling code here:
+        cardLayout.show(panelCard, "card1");
+    }                                        
+
+    private void jButton6ActionPerformed(java.awt.event.ActionEvent evt) {                                         
+        // TODO add your handling code here:
+        cardLayout.show(panelCard, "card3");
+    }                                        
+
+    private void jButton7ActionPerformed(java.awt.event.ActionEvent evt) {                                         
+        // TODO add your handling code here:
+        cardLayout.show(panelCard, "card5");
+    }                                        
+
 
     // Variables declaration - do not modify                     
-    private javax.swing.JLabel jLabel1;
-    private String personId;
+    private javax.swing.JPanel c1;
+    private javax.swing.JPanel c14;
+    private javax.swing.JPanel c15;
+    private javax.swing.JPanel c2;
+    private javax.swing.JPanel c3;
+    private javax.swing.JButton jButton3;
+    private javax.swing.JButton jButton4;
+    private javax.swing.JButton jButton5;
+    private javax.swing.JButton jButton6;
+    private javax.swing.JButton jButton7;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel12;
+    private javax.swing.JLabel jLabel13;
+    private javax.swing.JLabel jLabel14;
+    private javax.swing.JLabel jLabel15;
+    private javax.swing.JLabel jLabel51;
+    private javax.swing.JPanel jPanel1;
+    private javax.swing.JPanel jPanel3;
+    private javax.swing.JSeparator jSeparator4;
+    private javax.swing.JPanel panelCard;
     // End of variables declaration                   
 }
+
