+++ src/caroodj/Objects/CarDataFile.java
@@ -12,7 +12,7 @@ public class CarDataFile extends DataFile<Car>{
     }
 
     // Query database based on hashmap
-    public ArrayList<Car> queryDatabase(HashMap<String, String> query) {
+    public CarDataFile queryDatabase(HashMap<String, String> query) {
         Scanner dataFile = super.openFileRead();
         
         ArrayList<Car> carList = new ArrayList<Car>();
@@ -94,7 +94,8 @@ public class CarDataFile extends DataFile<Car>{
             }
         }
 
-        return carList;
+        this.lastQuery = carList;
+        return this;
     }
 
     public HashMap<String, String> createQuery(String[] car) {
